Initial pass at `editorPage.jsx`.
(cherry picked from commit 6cd56dfd6213011969937e962e0b0ca5d8109d97)
